# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tepati API
  version: 1.0.0
  summary: HTTP contract for the Tepati study-pact app.
  description: |
    Tepati keeps study promises. Two peers sign a *pact*: the *backer* funds a coin *pot*,
    and the *doer* submits daily *proof* before a cutoff. Coins are an IOU ledger. No real
    money moves through the app.

    This file is the contract. `docs/SPEC.md` is the source of truth for behaviour, and
    `docs/ARCHITECTURE.md §5` is the human-readable summary.

    ## Conventions

    - **Base path** `/api/v1` (see `servers`). Health checks live at the origin root.
    - **Auth** is a session cookie (`tepati_session`, HttpOnly). Every unsafe method also needs
      the `X-CSRF-Token` header, which the web app copies from the readable `tepati_csrf` cookie.
    - **Errors** are RFC 9457 `application/problem+json`. Clients branch on the stable `code`, never
      on `title` or `detail`. The status for each code is listed under `ErrorCode`.
    - **Authorisation.** A user who is not a member of a pact gets `404 not_found` for anything in
      it, so pact existence never leaks. A member who lacks the power for an action gets `403`.
    - **Idempotency.** Mutating endpoints accept an `Idempotency-Key` header. The first response
      is stored for 24 hours and replayed for the same user, route and key. Reusing a key with a
      different body returns `422 idempotency.key_reused`.
...
      pact's `coin_rate_idr`.
...
      (`null` on the last page).
...
servers:
...
security:
  - cookieAuth: []
    csrfToken: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create an account and start a session
      description: Sets the `tepati_session` and `tepati_csrf` cookies on success.
      security: []
      x-problem-codes: [validation.failed, auth.invalid_email, auth.invalid_name, auth.weak_password, auth.email_taken, auth.rate_limited]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Start a session
      description: |
        Sets the `tepati_session` and `tepati_csrf` cookies on success. Limited to 10 attempts per
        minute per IP and per email. A wrong email and a wrong password are indistinguishable.
      security: []
      x-problem-codes: [validation.failed, auth.invalid_credentials, auth.rate_limited]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
  /invites/{token}:
    parameters:
      - $ref: "#/components/parameters/InviteToken"
    get:
      tags: [pacts]
      operationId: previewInvite
      summary: Preview the terms behind an invite link
      description: |
        Public, because the invitee may not have an account yet. The token is the credential.
        It reveals only what the invitee needs to decide: the terms and who invited them.
      security: []
      x-problem-codes: [pact.invite_invalid, rate_limited]
      responses:
        "200":
          description: The invite preview.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitePreview" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tepati_session
      description: HttpOnly session cookie set by `register` and `login`.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit token, bound to the session. Required on every unsafe method, copied from the `tepati_csrf` cookie.
  parameters:
...
    InviteToken:
      name: token
      in: path
      required: true
      description: The one-time invite token from `proposePact`.
      schema: { type: string, minLength: 16, maxLength: 128 }
...
  responses:
    Problem:
      description: An error. Branch on `code`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error code. The web app maps each one to an i18n message.

        | Status | Codes |
        |---|---|
        | 400 | `validation.failed`, `auth.invalid_email`, `auth.invalid_name`, `auth.weak_password`, `pact.invalid_terms`, `pact.terms_members`, `proof.invalid_doc`, `checkin.reason_required`, `upload.size_mismatch` |
        | 401 | `auth.unauthenticated`, `auth.invalid_credentials` |
        | 403 | `auth.csrf`, `pact.not_backer`, `pact.not_doer`, `checkin.not_allowed` |
        | 404 | `not_found` (also for non-members) |
        | 405 | `method_not_allowed` |
        | 409 | `auth.email_taken`, `pact.invalid_state`, `pact.terms_mismatch`, `pact.member_missing`, `pact.limit_reached`, `checkin.invalid_transition`, `checkin.conflict`, `checkin.override_limit`, `checkin.rest_limit`, `upload.quota_exceeded`, `idempotency.in_progress` |
        | 410 | `pact.invite_invalid` |
        | 413 | `upload.too_large`, `request.too_large` |
        | 415 | `upload.unsupported_type`, `request.unsupported_media_type` |
        | 422 | `pact.signature_mismatch`, `checkin.deadline_passed`, `checkin.evidence_insufficient`, `idempotency.key_reused` |
        | 429 | `rate_limited`, `auth.rate_limited` |
        | 500 | `server.internal` |
        | 503 | `upload.queue_busy`, `server.unavailable` |
      enum:
        - validation.failed
        - not_found
        - method_not_allowed
        - request.too_large
        - request.unsupported_media_type
        - rate_limited
        - server.internal
        - server.unavailable
        - idempotency.in_progress
        - idempotency.key_reused
        - auth.unauthenticated
        - auth.csrf
        - auth.invalid_credentials
        - auth.email_taken
        - auth.invalid_email
        - auth.invalid_name
        - auth.weak_password
        - auth.rate_limited
        - pact.invalid_state
        - pact.invalid_terms
        - pact.terms_members
        - pact.terms_mismatch
        - pact.not_backer
        - pact.not_doer
        - pact.invite_invalid
        - pact.limit_reached
        - pact.signature_mismatch
        - pact.member_missing
        - proof.invalid_doc
        - checkin.invalid_transition
        - checkin.not_allowed
        - checkin.deadline_passed
        - checkin.reason_required
        - checkin.evidence_insufficient
        - checkin.override_limit
        - checkin.rest_limit
        - checkin.conflict
        - upload.too_large
        - upload.unsupported_type
        - upload.size_mismatch
        - upload.quota_exceeded
        - upload.queue_busy
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: JSON path of the offending field, for example `terms.members.<id>.penalty_per_miss`.
        message: { type: string }
    Problem:
      type: object
      description: RFC 9457 problem details with a stable `code`.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          description: A URI naming the problem type. `about:blank` when `code` says it all.
          default: about:blank
        title: { type: string, description: A short English summary. Not for display. }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: [string, "null"] }
        instance: { type: [string, "null"] }
        code: { $ref: "#/components/schemas/ErrorCode" }
        request_id: { type: [string, "null"], description: "Echo of the `X-Request-Id` header, for support." }
        errors:
          type: array
          description: Per-field detail, present for `validation.failed` and `pact.invalid_terms`.
          items: { $ref: "#/components/schemas/FieldError" }

    # ------------------------------------------------------------ users
    User:
      type: object
      required: [id, email, display_name, locale, timezone, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string, minLength: 1, maxLength: 80 }
        locale: { type: string, enum: [id, en] }
        timezone: { type: string, description: IANA zone, example: Asia/Jakarta }
        created_at: { type: string, format: date-time }
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 128, format: password }
        display_name:
          type: string
          minLength: 1
          maxLength: 80
          description: Also the signature the user types when accepting terms.
        locale: { type: string, enum: [id, en], default: id }
        timezone: { type: string, default: Asia/Jakarta, description: IANA zone. }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, maxLength: 128, format: password }
    MemberRef:
      type: object
      description: Just enough about a member to draw a line in their colour.
      required: [user_id, display_name, line_color]
      properties:
        user_id: { type: string, format: uuid }
        display_name: { type: string }
        line_color: { $ref: "#/components/schemas/LineColor" }
    LineColor:
      type: string
      pattern: "^#[0-9A-Fa-f]{6}$"
      description: The member's fixed colour for the whole pact.

    # ------------------------------------------------------------ terms and pacts
    Terms:
      type: object
      description: |
        The frozen rule set both members sign (SPEC §4). `terms_hash` is sha256 of its canonical JSON.
        The server validates every rule and answers `pact.invalid_terms` with per-field `errors`.
      required:
        - version
        - timezone
        - starts_on
        - ends_on
        - cutoff_local_time
        - grace_minutes
        - coin_rate_idr
        - initial_pot
        - pot_floor
        - pot_cap
        - review_window_hours
        - dispute_window_hours
        - dispute_resolution_hours
        - override_window_hours
        - max_overrides
        - backer_commits
        - members
      properties:
        version: { type: integer, minimum: 1, description: Schema version of the terms document. }
        timezone: { type: string, description: IANA zone, example: Asia/Jakarta }
        starts_on: { type: string, format: date }
        ends_on:
          type: string
          format: date
          description: Between 1 and 366 days after `starts_on`.
        cutoff_local_time:
          type: string
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          example: "23:59"
        grace_minutes: { type: integer, minimum: 0, maximum: 180 }
        coin_rate_idr: { type: integer, format: int64, minimum: 1, description: "Rupiah per coin. Display only, since the ledger is an IOU." }
        initial_pot: { type: integer, format: int64, minimum: 1 }
        pot_floor: { type: integer, format: int64, minimum: 0, description: The pot never drops below this. }
        pot_cap:
          type: [integer, "null"]
          format: int64
          description: Backer bonuses stop at this. `null` means no cap. Must be at least `initial_pot`.
        review_window_hours: { type: integer, minimum: 1, maximum: 72 }
        dispute_window_hours: { type: integer, minimum: 1, maximum: 72 }
        dispute_resolution_hours: { type: integer, minimum: 1, maximum: 72 }
        override_window_hours: { type: integer, minimum: 1, maximum: 72 }
        max_overrides: { type: integer, minimum: 0, maximum: 10, description: "0 disables the override power." }
        backer_commits: { type: boolean, description: Whether the backer also has a daily commitment. }
        members:
          type: object
          description: |
            Keyed by user id. Exactly the backer and one doer. Until the invitee joins, the doer
            slot is keyed by the nil UUID `00000000-0000-0000-0000-000000000000`.
          additionalProperties: { $ref: "#/components/schemas/MemberTerms" }
    MemberTerms:
      type: object
      required: [role, commitment, schedule, penalty_per_miss, rest_days, evidence]
      properties:
        role: { type: string, enum: [backer, doer] }
        commitment: { type: string, minLength: 1, maxLength: 200, description: Shown to the member every day. }
        schedule:
          type: array
          description: ISO weekdays, 1 = Monday. Empty means this member does not commit (backer only).
          uniqueItems: true
          maxItems: 7
          items: { type: integer, minimum: 1, maximum: 7 }
        penalty_per_miss:
          type: integer
          format: int64
          minimum: 1
          description: Coins taken from the pot when a doer misses, or added to it when the backer misses.
        rest_days: { type: integer, minimum: 0, description: Rest days this member may declare in the pact. }
        evidence: { $ref: "#/components/schemas/Evidence" }
    Evidence:
      type: object
      required: [min_attachments, min_words]
      properties:
        min_attachments: { type: integer, minimum: 0, maximum: 10 }
        min_words: { type: integer, minimum: 0 }
    PactStatus:
      type: string
      enum: [draft, proposed, scheduled, active, settling, completed, cancelled]
...
    InvitePreview:
      type: object
      required: [pact_title, inviter, terms, terms_hash, status, doer_slot_open]
      properties:
        pact_title: { type: string }
        inviter: { $ref: "#/components/schemas/MemberRef" }
        terms: { $ref: "#/components/schemas/Terms" }
        terms_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        status: { $ref: "#/components/schemas/PactStatus" }
        doer_slot_open: { type: boolean, description: False if someone already joined. }

    # ------------------------------------------------------------ check-ins
...